Fix space-group irrep generation for complex point-group irreps (#1876)
The `SpaceGroupBuilder._little_group_irreps` method has a bug, which
causes imaginary parts of the point-group irreps to be thrown away,
which leads to invalid space-group irreps if said point-group irreps are
complex. The existing tests didn't pick this up, as the full point
groups of the default lattices we ship have only real irreps.

This PR is a trivial fix of the bug, as well as an added test with
complex point-group irreps. The latter is a bit hacky, since the special
lattice constructors don't support changing the point group. This I will
fix in a follow-up PR after this is merged.

+++ netket/graph/space_group.py
@@ -233,7 +233,7 @@ class SpaceGroupBuilder:
         """
         idx = self._little_group_index(k)
         CT = self.little_group(k).character_table()
-        CT_full = np.zeros((CT.shape[0], len(self.point_group_)))
+        CT_full = np.zeros((CT.shape[0], len(self.point_group_)), dtype=CT.dtype)
         CT_full[:, idx] = CT
         return CT_full / idx.size if divide else CT_full
